# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Dashboard API",
    "description": "RESTful API for managing car rental operations including cars, bookings, customers, expenses, maintenance, and incidents.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token or API key authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "staff", "readonly"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "car_id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "make": { "type": "string" },
          "model": { "type": "string" },
          "customer_name": { "type": "string" },
          "customer_phone": { "type": "string", "nullable": true },
          "customer_id_passport": { "type": "string", "nullable": true },
          "customer_id": { "type": "integer", "nullable": true },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "status": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"] },
          "total_price": { "type": "number", "format": "float" },
          "deposit": { "type": "number", "format": "float" },
          "notes": { "type": "string", "nullable": true },
          "start_mileage": { "type": "integer", "nullable": true },
          "end_mileage": { "type": "integer", "nullable": true },
          "miles_driven": { "type": "integer" },
          "actual_return_date": { "type": "string", "format": "date", "nullable": true },
          "late_fee_amount": { "type": "number", "format": "float" },
          "created_at": { "type": "string", "format": "date-time" },
          "extras": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "extra_name": { "type": "string" },
                "extra_price": { "type": "number" }
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user and get JWT token",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "summary": "Get current authenticated user",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/api/bookings": {
      "get": {
        "summary": "List all bookings",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"] } },
          { "name": "car_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Booking" } }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new booking",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["car_id", "customer_name", "start_date", "end_date"],
                "properties": {
                  "car_id": { "type": "integer" },
                  "customer_name": { "type": "string" },
                  "customer_phone": { "type": "string", "nullable": true },
                  "customer_id_passport": { "type": "string", "nullable": true },
                  "customer_id": { "type": "integer", "nullable": true },
                  "start_date": { "type": "string", "format": "date" },
                  "end_date": { "type": "string", "format": "date" },
                  "status": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"], "default": "draft" },
                  "total_price": { "type": "number", "minimum": 0 },
                  "deposit": { "type": "number", "minimum": 0 },
                  "notes": { "type": "string", "nullable": true },
                  "extras": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "extra_name": { "type": "string" },
                        "extra_price": { "type": "number" }
                      }
                    }
                  },
                  "start_mileage": { "type": "integer", "nullable": true },
                  "end_mileage": { "type": "integer", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          },
          "400": { "description": "Invalid input or car not available" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          },
          "404": { "description": "Booking not found" }
        }
      },
      "put": {
        "summary": "Update booking",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "car_id": { "type": "integer" },
                  "customer_name": { "type": "string" },
                  "customer_phone": { "type": "string", "nullable": true },
                  "customer_id_passport": { "type": "string", "nullable": true },
                  "customer_id": { "type": "integer", "nullable": true },
                  "start_date": { "type": "string", "format": "date" },
                  "end_date": { "type": "string", "format": "date" },
                  "status": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"] },
                  "total_price": { "type": "number", "minimum": 0 },
                  "deposit": { "type": "number", "minimum": 0 },
                  "notes": { "type": "string", "nullable": true },
                  "extras": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "extra_name": { "type": "string" },
                        "extra_price": { "type": "number" }
                      }
                    }
                  },
                  "start_mileage": { "type": "integer", "nullable": true },
                  "end_mileage": { "type": "integer", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Booking not found" }
        }
      }
    },
    "/api/bookings/{id}/status": {
      "put": {
        "summary": "Update booking status",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["draft", "reserved", "confirmed", "active", "completed", "cancelled"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Booking" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Booking not found" }
        }
      }
    },
...
  },
...
}